requirements.txt
einops
numpy
onnx
onnxconverter-common
onnxruntime-gpu
opencv-python
torch
torchvision
pillow
